+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Use `hybrid-krompus` for integrated terminal in `VSCode`
 - Fixed `1password-gui` crashing when attempting to select a file
 - Removed `psi` and `tau`
 - Added `work` Firefox profile on `personal` machines